Move Billing module to KeilaCloud namespace
This module is not useful on self-hosted instances and specific to Keila Cloud.
