+++ config/test.exs
@@ -50,7 +50,7 @@ config :keila, Keila.Mailings.SenderAdapters,
 # Disable sending quotas by default in testing
 config :keila, Keila.Accounts, credits_enabled: false
 
-config :keila, Keila.Billing,
+config :keila, KeilaCloud.Billing,
   # Disable subscriptions
   enabled: false,
   # Disable paddle signature verification
