+++ extra/keila_cloud/config/cloud_config.exs
@@ -6,3 +6,9 @@ config :keila, Keila.Mailings.SenderAdapters,
   adapters: [
     KeilaCloud.Mailings.SendWithKeila | adapters
   ]
+
+config :keila, KeilaCloud.Billing,
+  # Disable subscriptions by default
+  enabled: false,
+  paddle_vendor: "2518",
+  paddle_environment: "sandbox"
