+++ lib/keila/mailings/send_welcome_email_worker.ex
@@ -28,15 +28,15 @@ defmodule Keila.Mailings.SendWelcomeEmailWorker do
     end
   end
 
-  defp ensure_feature_available(project_id) do
-    if Keila.Billing.feature_available?(project_id, :welcome_email) do
-      :ok
-    else
-      {:cancel, "Welcome email not enabled for account of project #{project_id}"}
-    end
-  end
-
   Keila.if_cloud do
+    defp ensure_feature_available(project_id) do
+      if KeilaCloud.Billing.feature_available?(project_id, :welcome_email) do
+        :ok
+      else
+        {:cancel, "Welcome email not enabled for account of project #{project_id}"}
+      end
+    end
+
     defp ensure_account_active(project_id) do
       case Keila.Accounts.get_project_account(project_id) do
         %{status: :active} -> :ok
@@ -44,6 +44,7 @@ defmodule Keila.Mailings.SendWelcomeEmailWorker do
       end
     end
   else
+    defp ensure_feature_available(_project_id), do: :ok
     defp ensure_account_active(_project_id), do: :ok
   end
