+++ lib/keila_web/controllers/account_controller.ex
@@ -1,8 +1,7 @@
 defmodule KeilaWeb.AccountController do
   use KeilaWeb, :controller
   import Ecto.Changeset
-  import Phoenix.LiveView.Controller
-  alias Keila.{Auth, Accounts, Billing}
+  alias Keila.{Auth, Accounts}
 
   @spec edit(Plug.Conn.t(), map()) :: Plug.Conn.t()
   def edit(conn, _) do
@@ -38,24 +37,12 @@ defmodule KeilaWeb.AccountController do
   defp render_edit(conn, changeset) do
     account = Accounts.get_user_account(conn.assigns.current_user.id)
     credits = if account, do: Accounts.get_credits(account.id)
-    subscription = if account, do: Billing.get_account_subscription(account.id)
-    plans = if Billing.billing_enabled?(), do: Billing.get_plans()
-    plan = if subscription, do: Billing.get_plan(subscription.paddle_plan_id)
 
     conn
     |> put_meta(:title, dgettext("auth", "Manage Account"))
     |> assign(:changeset, changeset)
     |> assign(:account, account)
     |> assign(:credits, credits)
-    |> assign(:subscription, subscription)
-    |> assign(:plans, plans)
-    |> assign(:plan, plan)
     |> render("edit.html")
-  end
-
-  def await_subscription(conn, _) do
-    live_render(conn, KeilaWeb.AwaitSubscriptionLive,
-      session: %{"current_user" => conn.assigns.current_user, "locale" => Gettext.get_locale()}
-    )
   end
 end
