+++ lib/keila_web/controllers/form_controller.ex
@@ -1,7 +1,8 @@
 defmodule KeilaWeb.FormController do
   use KeilaWeb, :controller
-  alias Keila.Contacts
+  require Keila
   import Phoenix.LiveView.Controller
+  alias Keila.Contacts
 
   plug :authorize when action not in [:index, :new, :delete, :display, :submit, :unsubscribe]
 
@@ -26,15 +27,20 @@ defmodule KeilaWeb.FormController do
   def edit(conn, _params) do
     current_project = current_project(conn)
 
-    double_opt_in_available? =
-      Keila.Billing.feature_available?(current_project.id, :double_opt_in)
+    Keila.if_cloud do
+      double_opt_in_available? =
+        KeilaCloud.Billing.feature_available?(current_project.id, :double_opt_in)
 
-    welcome_email_available? =
-      Keila.Billing.feature_available?(current_project.id, :welcome_email)
+      welcome_email_available? =
+        KeilaCloud.Billing.feature_available?(current_project.id, :welcome_email)
+    else
+      double_opt_in_available? = true
+      welcome_email_available? = true
+    end
 
     live_render(conn, KeilaWeb.FormEditLive,
       session: %{
-        "current_project" => current_project(conn),
+        "current_project" => current_project,
         "form" => conn.assigns.form,
         "locale" => Gettext.get_locale(),
         "double_opt_in_available" => double_opt_in_available?,
