+++ lib/keila_web/templates/account/edit.html.heex
@@ -1,154 +1,6 @@
 <div class="container mt-8" x-data x-init="Hooks.RememberUnsaved.mounted.call({el: $el})">
-  <%= if Keila.Billing.billing_enabled?() do %>
-    <div class="rounded shadow p-8 mt-8 max-w-5xl mx-auto flex flex-col gap-4 bg-gray-900 text-gray-50">
-      <h2 class="text-3xl font-bold">
-        <%= gettext("Subscription") %>
-      </h2>
-
-      <%= if  @subscription do %>
-        <%= case @subscription.status do %>
-          <% :active -> %>
-            <p class="flex gap-4 items-center">
-              <span class="flex h-2 w-8 items-center text-emerald-500">
-                <%= render_icon(:check) %>
-              </span>
-              <%= gettext(
-                "You currently have an active subscription. Thanks for supporting Keila!"
-              ) %>
-            </p>
-          <% :paused -> %>
-            <p class="flex gap-4 items-center">
-              <span class="flex h-2 w-8 items-center text-red-500">
-                <%= render_icon(:information_circle) %>
-              </span>
-              <%= gettext(
-                "Your subscription is currently suspended. Please update your payment method to continue using Keila."
-              ) %>
-            </p>
-          <% :past_due -> %>
-            <p class="flex gap-4 items-center">
-              <span class="flex h-2 w-8 items-center text-red-500">
-                <%= render_icon(:information_circle) %>
-              </span>
-              <%= gettext(
-                "There was an error processing your payment. Please update your payment method to continue using Keila."
-              ) %>
-            </p>
-          <% :deleted -> %>
-            <p class="flex gap-4 items-center">
-              <span class="flex h-2 w-8 items-center text-red-500">
-                <%= render_icon(:information_circle) %>
-              </span>
-              <%= gettext(
-                "Your subscription has been cancelled. Please contact us if you want to start a new subscription."
-              ) %>
-            </p>
-        <% end %>
-
-        <%= if @subscription.status != :deleted do %>
-          <div class="flex flex-row">
-            <div class="text-center flex flex-col gap-4 items-center grow-0">
-              <h3 class="font-bold">
-                <span class="text-2xl">
-                  <%= @plan.name %>
-                </span>
-                <span class="block -mt-1 text-sm text-emerald-100">
-                  <%= raw(
-                    gettext("%{price} / month",
-                      price:
-                        ~s{<span class="paddle-net" data-product="#{@plan.paddle_id}"></span>}
-                    )
-                  ) %>
-                </span>
-                <span class="block text-xs text-emerald-300">
-                  <%= raw(
-                    gettext("(%{price} incl. tax)",
-                      price:
-                        ~s{<span class="paddle-gross" data-product="#{@plan.paddle_id}"></span>}
-                    )
-                  ) %>
-                </span>
-              </h3>
-              <ul>
-                <li class="flex">
-                  <check-complete class="mr-2 shrink-0" />
-                  <%= gettext("%{monthly_emails} emails/month",
-                    monthly_emails: @plan.monthly_credits
-                  ) %>
-                </li>
-                <li class="flex">
-                  <check-complete class="mr-2 shrink-0" />
-                  <%= gettext("unlimited contacts") %>
-                </li>
-              </ul>
-              <div class="flex gap-4">
-                <%= if @subscription.update_url do %>
-                  <a class="button" target="_blank" href={@subscription.update_url}>
-                    <%= gettext("Update payment method") %>
-                  </a>
-                <% end %>
-                <%= if @subscription.cancel_url do %>
-                  <a class="button" target="_blank" href={@subscription.cancel_url}>
-                    <%= gettext("Cancel subscription") %>
-                  </a>
-                <% end %>
-              </div>
-            </div>
-          </div>
-        <% end %>
-      <% else %>
-        <%= gettext("Subscribe now and start sending emails!") %>
-
-        <div class="grid grid-cols-3 gap-8">
-          <%= for plan <- @plans do %>
-            <div class="p-4 text-center flex flex-col gap-4 items-center rounded-sm">
-              <h3 class="font-bold">
-                <span class="text-2xl">
-                  <%= plan.name %>
-                </span>
-                <span class="block -mt-1 text-sm text-emerald-100">
-                  <%= raw(
-                    gettext("%{price} / month",
-                      price: ~s{<span class="paddle-net" data-product="#{plan.paddle_id}"></span>}
-                    )
-                  ) %>
-                </span>
-                <span class="block text-xs text-emerald-300">
-                  <%= raw(
-                    gettext("(%{price} incl. tax)",
-                      price:
-                        ~s{<span class="paddle-gross" data-product="#{plan.paddle_id}"></span>}
-                    )
-                  ) %>
-                </span>
-              </h3>
-              <ul>
-                <li class="flex">
-                  <check-complete class="mr-2 shrink-0" />
-                  <%= gettext("%{monthly_emails} emails/month",
-                    monthly_emails: plan.monthly_credits
-                  ) %>
-                </li>
-                <li class="flex">
-                  <check-complete class="mr-2 shrink-0" />
-                  <%= gettext("unlimited contacts") %>
-                </li>
-              </ul>
-              <button
-                class="paddle_button button button--large button--cta w-full block justify-center"
-                data-product={plan.paddle_id}
-                data-email={@current_user.email}
-                data-passthrough={Jason.encode!(%{"account_id" => @account.id})}
-                data-theme="none"
-                data-success={Routes.account_url(@conn, :await_subscription)}
-              >
-                <%= gettext("Get started") %>
-              </button>
-            </div>
-          <% end %>
-        </div>
-      <% end %>
-    </div>
+  <%= Keila.if_cloud do %>
+    <.subscription_status account={@account} current_user={@current_user} />
   <% end %>
 
   <%= if Keila.Accounts.credits_enabled?() do %>
@@ -238,10 +90,3 @@
     </div>
   </.form>
 </div>
-
-<script src="https://cdn.paddle.com/paddle/paddle.js">
-</script>
-<script type="text/javascript">
-  Paddle.Environment.set('<%= Application.get_env(:keila, Keila.Billing) |> Keyword.fetch!(:paddle_environment) %>')
-  Paddle.Setup({ vendor: <%= Application.get_env(:keila, Keila.Billing) |> Keyword.fetch!(:paddle_vendor) %> })
-</script>
