+++ lib/keila_web/views/account_view.ex
@@ -1,3 +1,8 @@
 defmodule KeilaWeb.AccountView do
+  require Keila
   use KeilaWeb, :view
+
+  Keila.if_cloud do
+    import KeilaCloud.Components.SubscriptionStatus
+  end
 end
